modularize and set up preview for amendments
