+++ routes/mindep_router.py
@@ -54,6 +54,7 @@ def get_contents_of_amendment_gazette(gazette_number: str, date: str):
     try:
         data = utils.load_mindep_gazette_data_from_JSON(gazette_number, date)
         transactions = mindep_gazette_processor.process_amendment_gazette(gazette_number, date, data)
+        trans_database.create_record(gazette_number,"mindep","amendment", date)
         return transactions
     except FileNotFoundError:
         return {"error": f"Gazette file for {gazette_number}, {date} not found."}
